+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- New `rebar3 grisp report` command: groups and packs usefull debug information
+about the current grisp project
+
 ## [2.4.0] - 2022-07-18
 
 ### Added